pyproject.toml
[project]
name = "web-collector"
version = "0.1.0"
description = "A Telegram bot for analyzing website templates and finding similar websites"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "python-telegram-bot>=20.7",
    "playwright>=1.40.0",
    "aiosqlite>=0.19.0",
    "beautifulsoup4>=4.12.0",
    "scikit-learn>=1.3.0",
    "pyyaml>=6.0.1",
    "numpy>=1.24.0",
    "typing-extensions>=4.8.0",
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.9.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
]
lint = [
    "black>=23.9.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src --cov-report=term-missing"

[tool.black]
line-length = 100
target-version = ["py39"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = [
    "playwright.*",
    "sklearn.*",
]
ignore_missing_imports = true

[tool.pdm.scripts]
test = "pytest"
lint = { composite = ["black .", "isort .", "mypy src tests"] }
format = { composite = ["black .", "isort ."] }
start = "python main.py"
